reduce MSVC warnings a bit

+++ lib/jxl/enc_external_image.cc
@@ -137,7 +137,7 @@ Status ConvertFromExternal(Span<const uint8_t> bytes, size_t xsize,
   if (float_in) {
     RunOnPool(
         pool, 0, static_cast<uint32_t>(ysize), ThreadPool::SkipInit(),
-        [&](const int task, int /*thread*/) {
+        [&](const uint32_t task, size_t /*thread*/) {
           const size_t y = task;
           size_t i = row_size * task;
           float* JXL_RESTRICT row_out = channel->Row(y);
@@ -172,7 +172,7 @@ Status ConvertFromExternal(Span<const uint8_t> bytes, size_t xsize,
     float mul = 1. / ((1ull << bits_per_sample) - 1);
     RunOnPool(
         pool, 0, static_cast<uint32_t>(ysize), ThreadPool::SkipInit(),
-        [&](const int task, int /*thread*/) {
+        [&](const uint32_t task, size_t /*thread*/) {
           const size_t y = task;
           size_t i = row_size * task;
           float* JXL_RESTRICT row_out = channel->Row(y);
@@ -251,7 +251,7 @@ Status ConvertFromExternal(Span<const uint8_t> bytes, size_t xsize,
     for (size_t c = 0; c < color_channels; ++c) {
       RunOnPool(
           pool, 0, static_cast<uint32_t>(ysize), ThreadPool::SkipInit(),
-          [&](const int task, int /*thread*/) {
+          [&](const uint32_t task, size_t /*thread*/) {
             const size_t y = get_y(task);
             size_t i =
                 row_size * task + (c * bits_per_sample / jxl::kBitsPerByte);
@@ -290,7 +290,7 @@ Status ConvertFromExternal(Span<const uint8_t> bytes, size_t xsize,
     for (size_t c = 0; c < color_channels; ++c) {
       RunOnPool(
           pool, 0, static_cast<uint32_t>(ysize), ThreadPool::SkipInit(),
-          [&](const int task, int /*thread*/) {
+          [&](const uint32_t task, size_t /*thread*/) {
             const size_t y = get_y(task);
             size_t i = row_size * task + c * bytes_per_channel;
             float* JXL_RESTRICT row_out = color.PlaneRow(c, y);
@@ -335,7 +335,7 @@ Status ConvertFromExternal(Span<const uint8_t> bytes, size_t xsize,
     if (float_in) {
       RunOnPool(
           pool, 0, static_cast<uint32_t>(ysize), ThreadPool::SkipInit(),
-          [&](const int task, int /*thread*/) {
+          [&](const uint32_t task, size_t /*thread*/) {
             const size_t y = get_y(task);
             size_t i = row_size * task +
                        (color_channels * bits_per_sample / jxl::kBitsPerByte);
@@ -371,7 +371,7 @@ Status ConvertFromExternal(Span<const uint8_t> bytes, size_t xsize,
       float mul = 1. / ((1ull << bits_per_sample) - 1);
       RunOnPool(
           pool, 0, static_cast<uint32_t>(ysize), ThreadPool::SkipInit(),
-          [&](const int task, int /*thread*/) {
+          [&](const uint32_t task, size_t /*thread*/) {
             const size_t y = get_y(task);
             size_t i = row_size * task + color_channels * bytes_per_channel;
             float* JXL_RESTRICT row_out = alpha.Row(y);
